将actionType修改为int64类型，增加redis操作

+++ controller/favorite.go
@@ -17,7 +17,7 @@ type FavoriteController struct{}
 type FavoriteActionRequest struct {
 	Token      string `query:"token"              vd:"len($)>0"`   // 用户鉴权token
 	VideoID    int64  `query:"video_id,string"    vd:"$>0"`        // 视频id
-	ActionType int    `query:"action_type,string" vd:"$==1||$==2"` // 1-点赞，2-取消点赞
+	ActionType int64    `query:"action_type,string" vd:"$==1||$==2"` // 1-点赞，2-取消点赞
 }
 
 type FavoriteListRequest struct {
